if we don't have all the packages on headless init, wait for a frame
we notice this issue in `#'kotct/check-dependency-list` and throw
`'daemon-mode` back to `init.el`. we then stop what we're doing and
put everything on the `after-make-frame-functions` hook. when the
frame is created, we prompt about installing missing packages and
finish loading.

haven't tested yet with if user configs are missing packages at a
headless startup. that's next.