Adjust versioning and release scheme
This is somewhat more compliant with GitLab flow, though the notable
problem with this is that `stable` will contain version bump commits
that `master` does not have.  Subsequent releases may need to get
merged back into `master` so that we inform git that our `-pre`
versions are authoritative.

But this is a good first pass.

And since this is a scheme that I can get behind,

[GitHub] This commit resolves #57.