Add the rest of the specs for the sh configuration
Note that I did not include a test for `sh-mode-hook` having `indent-tabs-mode` set to `t` by default.  This is primarily due to concerns about equality checking, since it currently runs a `lambda`.  I plan to make API changes to `indentation` to improve all of this, however, so it might be best to ship it like this.