Add `force-load` optional argument to `kotct/switch-to-theme`
This will be used to instruct `kotct/switch-to-theme` to load the file
without worrying about whether or not it is already known.

For performance reasons, this is optional, (and will be nil in all
instances where it is already used) because loading a theme over and
over again requires evaluation.