Add support for display-line-numbers on Emacs >= 26.1
Here we add a helper function, kotct/display-line-numbers--turn-off
which can be used in place of lambdas in hooks to turn stuff
off. (From a cursory search I was unable to find an equivalent to
linum-off for display-line-numbers.)

Also, add a helper method which disables global-linum-mode and enables
global-display-line-numbers-mode.  I was considering simply turning
this on in a prog-mode hook, but this seems better.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>

[GitHub] This commit resolves #98.