init: Remove unnecessary call to package-initialize
According to Emacs 26.x+, this is not necessary.  It certainly seems
strange, but making this change does not affect the results of startup
at all for me, besides shaving just a little bit of time off.

Since we `require' every ELPA package before we configure it, this
shouldn't have any danger of affecting reachability, unless that
assumption is violated anywhere.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>

[GitHub] This commit resolves #123.