init: Move most requires into eval-when-compile blocks
Symbols typically fall into two categories: those registered explicitly
with `autoload` (e.g. via `;;;###autoload` and such), and those referred
to via the `register-definition-prefixes` function which helps to inform
autoloading.  It seems that those in the latter category will sometimes
fail to be detected by the system.  A third---and quite pesky---category
is those which do not have any autoload definitions prepared for them,
nor any prefixes hinting where autoload can hunt for their definitions.

Now, all external require statements are performed according to a
somewhat straightforward set of rules:

- If a (require) *is* directly used at evaluation time, it must be
  required immediately.

- If a (require) is not needed at evaluation time, (i.e. the file being
  evaluated never refers to any symbols present there) then the require
  can safely be moved into an eval-when-compile.

In some cases, moving a require into an eval-when-compile makes the
compiler a bit concerned, so manual (autoload) invocations seem to
effectively inform the compiler that yes, indeed, symbols are to be
autoloaded.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>