startup: Add a printout for the number of GCs done during startup
Now, we can get some idea for what is happening when Emacs is starting
up.  Here we don't actually print any information about the *time* spent
in GC, but we do get an idea for the number of GCs done.

The Emacs garbage collector operates by finding and "marking" all Lisp
objects that are still accessible to Lisp programs.  It assumes that all
symbols and their associated values and function definitions, and things
present on the stack are accessible.  Then any objects reachable through
these presumed-accessible objects are also considered accessible.  After
the process of markinging accessible Lisp objects is done, those left
unmarked are garbage, and cannot be reached by normal means.

It can be quite helpful to know how many GC operationss were performed
during startup to see if the GC parameters are set properly.  On my
system, for example, the gc-cons-threshold is merely 800_000 bytes (800
KB).  This is the default for 64-bit systems, but results in 42 GC
sweeps being performed during startup, which take a non-trivial amount
of time.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>