ci: Overhaul undercover file location
Previously, the ~/.emacs.d/lisp/ directory was consulted, but this
required the code to be installed into that directory and the script
file to be executed via that path.  (This meant that executing the
script directly instead of via the installed symlink to ~/.emacs.d/
would not work.)

Now, we simply expand all files and pass that as an argument list to
the undercover entrypoint.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>